+++ src/interfaces/index.ts
@@ -1,4 +1,6 @@
-export interface AppStore {
-  text: string;
+export interface Todos {
+  userId?:number;
+  title: string;
   completed: boolean;
-}[];
+  id: number;
+};
